Move Formation CRD into a separate package to cut dependency
