+++ cmd/crd/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     deps = [
         "//cmd:go_default_library",
         "//pkg/composition/crd:go_default_library",
-        "//pkg/formation:go_default_library",
+        "//pkg/formation/crd:go_default_library",
         "//pkg/ops:go_default_library",
         "//pkg/orchestration:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/resources:go_default_library",
