+++ pkg/formation/BUILD.bazel
@@ -2,16 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "client.go",
-        "controller.go",
-    ],
+    srcs = ["controller.go"],
     importpath = "github.com/atlassian/voyager/pkg/formation",
     visibility = ["//visibility:public"],
     deps = [
         "//:go_default_library",
-        "//pkg/api/schema:go_default_library",
-        "//pkg/apis/formation:go_default_library",
         "//pkg/apis/formation/v1:go_default_library",
         "//pkg/apis/orchestration/v1:go_default_library",
         "//pkg/formation/client/typed/formation/v1:go_default_library",
@@ -28,7 +23,6 @@ go_library(
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
         "//vendor/go.uber.org/zap:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
-        "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
