+++ pkg/formation/crd/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["crd.go"],
+    importpath = "github.com/atlassian/voyager/pkg/formation/crd",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/api/schema:go_default_library",
+        "//pkg/apis/formation:go_default_library",
+        "//pkg/apis/formation/v1:go_default_library",
+        "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+    ],
+)
